Update HomeController.cs
This is how I usually handle nested if statements. It is cleaner that way in my opinion. You might not need the else statement either, I don't really know if the Response.Redirect method will terminate or not thats why I used an else statement just in case. Took me a while to get into a habbit like this just some advice the project looks great by the way.